+++ db/sqlite/include/sqlite_result.hpp
@@ -65,7 +65,7 @@ protected:
   virtual void read(const char *id, bool &x);
   virtual void read(const char *id, float &x);
   virtual void read(const char *id, double &x);
-  virtual void read(const char *id, char *x, int s);
+  virtual void read(const char *id, char *x, size_t s);
   virtual void read(const char *id, varchar_base &x);
   virtual void read(const char *id, std::string &x);
   virtual void read(const char *id, oos::date &x);
